+++ test_manager/utils/test_utils.go
@@ -1,4 +1,4 @@
-package test_utils
+package utils
 
 import (
 	"context"
